requirements.txt
Flask>=2.0
srt>=3.5
MeCab-python3>=1.0
